requirements-assignment2.txt
torch
torchvision
tqdm
fastapi
uvicorn
pillow
python-multipart
